requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
supabase>=2.0.0
python-telegram-bot>=20.7
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
httpx>=0.25.0
google-generativeai>=0.3.0
openai>=1.0.0